add a cheesy heap allocator implementation to unlock more data structures
A pretty awful and very forced port of the cortex-m linked list
allocator. The main issue is that the RiscV runtime doesn't seem
to support atomic boolean swaps, which means the Mutex implementation
won't work.

Pulled in a Mutex implementation and lied to the system that the
mutex is always available. This will get us through initial board
bring-up but someone with real OS knowledge needs to fix this
down the road....