Wrap BetrustedDisplay in Mutex; fix issue #1
Issue #1 -- the problem with the dirty bit is that the LCD
region was originally mapped to 0x5000_0000, which is cacheable.
This is not the right behavior; mapping to 0xB000_0000 (a non-
cacheable region) fixes the dirty bit behavior.

wraping BetrustedDisplay in a Mutex allows the display object
to be borrowed/copied throughout the program without having
to re-allocate it inside e.g. loops.

Some concerns about performance wrt to the Mutex locking as
things scale up, need to be smart at the code level about
lock() calls and memoizing things like the result of the
.size() in the code.