page read burst length of 8 fails; length of 4 works
In theory, the datasheet says that page reads should work across
a 16-word span.

However, in testing in C models, we could only get reliable
operations with an 8-word span. Now with Rust running, we
get page read failures with 8 words. Reducing that width down
to 4 words, however, seems to get page reads working again.

Not sure what the cause is of the discrepancy, but bursting
on page reads will improve cache fill times by 2x so we should
retain this feature and mark it as a curio to debug later on.