slow down COM SPI clock to 20 MHz
The UP5K measured Tpd performance is around 18ns, which
is just a hair too slow to meet timing at 24MHz.

Reducing clock to 20MHz opens up ~4ns margin for PVT
note (50ns - 41.7ns)/2 = ~4ns (the /2 because we're doing
rising to falling edge capture).

The Spartan side should be pretty robust, but the timing
on the UP5K might vary a bit depending upon the routing,
it's not clear to me that the Yosys router really closes
timing that well...may need even more margin if the link
isn't stable!