refactor the SPI master interface
Synchronization with the SPI interface is a bit challenging.
In the best of cases, the CPU runs so fast it can read back
SPI status registers before they have a chance to update (CPU
runs at 5x SPI clock). In the worst of cases, the CPU hits
a cache fault and runs so slowly a whole SPI transaction can
finish before the CPU recovers.

This is compounded by the fact that the CSR bus is non-atomic
for data widths larger than a byte.

Add a two-phase handshake, which consists of "clear done",
(load Tx and go), and "wait done". This hand shake is a little
inefficient but it ensures that the CPU is synced with the
interface on polled implementations whether running fast or slow.

Interrupts of course will solve much of this.