add COM module beginnings
COM is the module that is Betrusted's interface to the insecure
world. It's the sole I/O port. It's also just a simple SPI master
interface right now, with very few trappings around it.

The basic COM protocol at the moment sends a command to the EC
and expects data in return. For example, writing the value
0x8000 to the EC tells it to return the value of the charger
registers on the next 8 successive COM transactions.

The EC has trouble keeping up with the COM bus. We're going to have
to noodle on how to improve this.