modify fb call to be non-blocking using a polled method
not sure if this should be applied to the final OS, because
the RAM area for the LCD FB is very fast. But, make it so that
calls to flush() succeed immediately, and only do anything if
more 25ms have passed and the LCD is not currently processing
an update.

This is done by creating a backing framebuffer where all the
rendering happens, and then a very fast copy of the entire
framebuffer to the LCD is pushed so that the call can
succeed quickly.