add patch to fix VADC_N issue
apparently, even the _N inputs to the XADC has to be broken out, even
if used in "unipolar" mode. :( Thought I could tie these to ground
internally, but that actually leaves the node floating. :-/

this will need a board spin to fix properly, but this patch for now
works. The hardware side needs a cut on NOISEBIAS_ON, tying that to
3.3V (so it's always on) and then taking the remnant stub to the FPGA
and tying it to GND. It turns out that somehow magically this also
fixes GND for the other signal paths even though they aren't wired up.
I suspect this is just due to remnant charge on the S/H cap, so
in fact the floating channels have degraded performance, but it
seems "good enough" to get us moving.