add the key store ROM as hard-located LUT primitives
These should always be compiled in so that the impact of the
location of the hard-located key ROM can be assessed as the
design evolves. The connectivity to the ROM can change, but
the location and pin ordering of the LUTs may not.

Now during compilaton, a rom.db file is output, which is a database
that is used by downstream scripts to generate the patching code
for the bitstream to repopulate the ROM with the cryptographic
key material as it is generated.

Timing analysis shows that the hard-coding of the position of
the ROM has minimal impact on the performance of the design, but
as the design starts to become more crowded it may become
advantageous to relocate the ROM. This is especally true as
eventually the ROM will migrate from being a CSR-accessible
block to a block that is only readable by a crypto accelerator.

If a migration happens, the rom.db file will need to be fed
into the associated scripts to generate the bitstream patching
code. As the patching has to happen on the fly in a memory-limited
environment, the patching code is derived from generic databases,
but mapped to a set of hard-coded offsets. Since the location
of the ROM is fully deterministic in the bitstream (barring an
intentional migration by the designer), this is an efficient
method for inserting the self-generated key material into the
bitstream.

In other words: we can move that ROM around as we define and
evolve the SoC, but once we go to production, moving that ROM
will involve coordinate a new set of positions with the new
locations in the bitstream, which is a versioning nightmare.