fix output LUT bug, ordering of blocks; start on automation
The mapping of LUT bits to key ROM bits works as expected, but
the mapping of LUTs to higher-order bits is a little weird.
I think this is because the CLB order is not striped nicely
through the bitstream, but that itself is also permuted relative
to the numbering of the slices.

This is fixed up by fixing up the mapping of the high two bits
of the address to the lut selectors. I don't think it'll change
due to any weird python reordering of dictionary to keyorder,
but something to keep in mind.