prove out a PHY schema that can meet OPI timing @ 100MHz
The only timing escape at the moment is when we fall back to
SPI mode -- the SPI->FPGA path is "slow" because the IDDR
is already allocated to OPI mode sampling (and the clock
to that IDDR is hard-wired to the DQS input), and it
must reach instead a fabric FDRE to sample the data. This
full path takes a bit over 5ns, and the spec'd delay of
the FLASH memory is 5ns as well, so for a 10ns clock
period, we find ourselves in trouble.

Timing misses by 0.5ns -- /probably/ this is OK, because
in the analysis path there's about 0.25ns of margin built-in,
and this is done using worst-case "slow" parameters on all
items in between. Considering the specs are -40C-85C on the
ROM, 0-100C for the FPGA, and I have never seen the design
run hotter than ~40C, I think there is a good chance
we can "cheat" and spec the input delay to 4.5ns and expect
the design will be reliable. If anything, the hold time (fast
path) is probably more a worry than the setup time (slow path).