SPI interface at 100MHZ SDR working
Requires rework of cutting POWER_S1 (L13), patching it to CCLK,
and then tying the former POWER_S1 pin to 1.8V_T.

The hardware patch is necessary because feeding the clock
back out the CCLK port can't meet timing -- the CCLK out port
has a timing variation of 0.5-7.5ns, which is greater than
a single data period at 100MHz DDR (that is, 5ns). By
patching to an actual GPIO and forcing CCLK to tri-state,
we can use an ODDR2 primitive to drive the clock and achieve
sub-nanosecond level timing closure between the CLK and data
pins going to the SPI ROM.

Also modify the simulation to boot from SPI, so we can
do an end-to-end simulation of the boot process.