checkpoint of code before total refactor of DOPI machine
decided that the current approach to OPI (two PHYs alternately
managed by one MAC) is not going to work. The timing between
SPI and DOPI is too different, plus, it turns out it is not
possible to interleave SPI and DOPI formats, so we're going
to refactor this interface into two clean, different interfaces
so I don't have to constantly step around SPI assumptions in
DOPI mode.